Refactoring av saksflyt/unntaksperiode => saksflyt/unntaksperioder OG nytt endepunkt [PUT] saksflyt/unntaksperioder/anmodning
