+++ scripts/modules/saksflyt/index.js
@@ -1,7 +1,7 @@
 const { endreperiode, fattet } = require('./vedtak');
-const { anmodning, godkjenn, innhentinfo, ikkegodkjenn } = require('./unntaksperiode');
+const { anmodning, godkjenn, innhentinfo, ikkegodkjenn } = require('./unntaksperioder');
 
 module.exports = {
   vedtak: { endreperiode, fattet },
-  unntaksperiode: { anmodning, godkjenn, innhentinfo, ikkegodkjenn }
+  unntaksperioder: { anmodning, godkjenn, innhentinfo, ikkegodkjenn }
 };
